+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Upgrade to MechanicalSoup 0.10.0
 - Improve whitelabel capabilities
 - Improve error response handling for IFI Claims
+- Set default language for "dpmaregister" crawler to English
 
 2018-01-22 0.148.1
 ==================